Add stub for test-harness and OS-specific setup
CMakeLists.txt has been modified to detect (on macOS) whether your macOS
version is >= 10.15. If it is, then the test harness will compile with the
native implementation of <filesystem>. Otherwise, it will use a library for
compatibility.

This project now uses submodules! To get the submodules, run:
```
git submodule update --init
```